+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+### Changed
+
+- [BREAKING]: `Throttle::fuzzy_fn()` & `ThrottlePool::fuzzy_fn()` now accept
+  `fn(Duration) -> Duration` rather than original `Option<fn(Duration) -> Duration>`.
+- [BREAKING]: set `fuzzy_fns` feature as optional.
+
+
 
 ## [0.1.1] - 2020-11-28